requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic<2.0.0
python-dotenv==1.0.0
email-validator==2.2.0